mulDiv returns 0 if numerator is 0 and denominator isn't: * Otherwise overflow checks result in divide by 0. * If denominator is 0, let the divide by 0 exception throw. * Move mulDiv out of STAmount
