+++ src/ripple/protocol/impl/STAmount.cpp
@@ -1275,40 +1275,6 @@ divRound (STAmount const& num, STAmount const& den,
     return result;
 }
 
-// compute (value)*(mul)/(div) - avoid overflow but keep precision
-std::uint64_t
-mulDiv(std::uint64_t value, std::uint64_t mul, std::uint64_t div)
-{
-    lowestTerms(value, div);
-    lowestTerms(mul, div);
-
-    if (value < mul)
-        std::swap(value, mul);
-    const auto max = std::numeric_limits<std::uint64_t>::max();
-    if (value > max / mul)
-    {
-        value /= div;
-        if (value > max / mul)
-            Throw<std::overflow_error> ("mulDiv");
-        return value * mul;
-    }
-    return value * mul / div;
-}
-
-// compute (value)*(mul)/(div) - avoid overflow but keep precision
-std::uint64_t
-mulDivNoThrow(std::uint64_t value, std::uint64_t mul, std::uint64_t div)
-{
-    try
-    {
-        return mulDiv(value, mul, div);
-    }
-    catch (std::overflow_error)
-    {
-        return std::numeric_limits<std::uint64_t>::max();
-    }
-}
-
 std::uint32_t
 STAmountCalcSwitchovers::enableUnderflowFixCloseTime ()
 {
