+++ src/ripple/unity/basics.cpp
@@ -25,6 +25,7 @@
 #include <ripple/basics/impl/CountedObject.cpp>
 #include <ripple/basics/impl/Log.cpp>
 #include <ripple/basics/impl/make_SSLContext.cpp>
+#include <ripple/basics/impl/mulDiv.cpp>
 #include <ripple/basics/impl/RangeSet.cpp>
 #include <ripple/basics/impl/ResolverAsio.cpp>
 #include <ripple/basics/impl/strHex.cpp>
